+++ inst/doc/multidog.R
@@ -7,6 +7,7 @@ knitr::opts_chunk$set(
 )
 
 ## ----setup--------------------------------------------------------------------
+library(future)
 library(updog)
 data("uitdewilligen")
 
@@ -20,7 +21,7 @@ setdiff(colnames(sizemat), colnames(refmat))
 setdiff(rownames(sizemat), rownames(refmat))
 
 ## -----------------------------------------------------------------------------
-parallel::detectCores()
+future::availableCores()
 
 ## -----------------------------------------------------------------------------
 mout <- multidog(refmat = refmat, 
@@ -28,6 +29,20 @@ mout <- multidog(refmat = refmat,
                  ploidy = ploidy, 
                  model = "norm",
                  nc = 2)
+
+## ---- eval = FALSE------------------------------------------------------------
+#  future::plan(future::multisession, workers = nc)
+
+## ---- eval = FALSE------------------------------------------------------------
+#  future::plan(future::multicore, workers = 2)
+#  mout <- multidog(refmat = refmat,
+#                   sizemat = sizemat,
+#                   ploidy = ploidy,
+#                   model = "norm",
+#                   nc = NA)
+#  
+#  ## Shut down parallel workers
+#  future::plan(future::sequential)
 
 ## -----------------------------------------------------------------------------
 plot(mout, indices = c(1, 5, 100))
@@ -48,4 +63,9 @@ dim(mout$inddf)
 mout_cleaned <- filter_snp(mout, prop_mis < 0.05 & bias > exp(-1) & bias < exp(1))
 dim(mout_cleaned$snpdf)
 dim(mout_cleaned$inddf)
+
+## ---- eval = FALSE------------------------------------------------------------
+#  # install.packages("BiocManager")
+#  # BiocManager::install(c("VariantAnnotation", "GenomicRanges", "S4Vectors", "IRanges"))
+#  export_vcf(obj = mout, filename = "./multidog_fit.vcf")