+++ NEWS.md
@@ -1,3 +1,17 @@
+# updog 2.1.1
+
+1. Added an upper bound to the sequencing error rate in `flexdog_full()` (and, hence, `flexdog()` and `multidog()`). This protects against some poor behavior observed in a corner case. Specifically, F1 populations where the offspring are all the same genotype and is sequenced at moderate to low depth.
+2. Fixed some stale URLs, fixed some style issues found by lintr.
+
+# updog 2.1.0
+
+1. The parallel backend in `multidog()` is now handled by the [`future`](https://cran.r-project.org/package=future) package. If you use the `nc` argument in `multidog()`, it should still run in parallel using multiple R sessions on your local machine. However, you can now use the functionality of `future` to choose your own evaluation strategy, after setting `nc = NA`. This will also allow you to use schedulers in high performance computing environments through the [`future.batchtools`](https://cran.r-project.org/package=future.batchtools) package. See the `multidog()` function documentation for more details.
+2. The vignette "Genotyping Many SNPs with multidog()" goes through an example of using the `future` package.
+3. A new experimental function, `export_vcf()`, is in the works to export `multidog` objects to a VCF file. This is not yet exported because I still have a few bugs to fix.
+4. `plot.multidog()` will now plot the parent read-counts in F1 and S1 populations.
+5. Internally, `multidog()` now uses iterators through the [`iterators`](https://cran.r-project.org/package=iterators) package to send only subsets of the data to each R process.
+6. A new internal `.combine` function is used in the `foreach()` call of `multidog()` in order to decrease the memory usage of `multidog()`.
+
 # updog 2.0.2
 
 This is a massive edit of the updog software. Major changes include:
@@ -27,7 +41,7 @@ This is a massive edit of the updog software. Major changes include:
 
 # updog 1.2.0
 
-- Adds `multidog()` for genotypying multiple SNPs using parallel computing.
+- Adds `multidog()` for genotyping multiple SNPs using parallel computing.
 - Adds `plot.multidog()` for plotting the output of `multidog()`.
 - Adds `format_multidog()` for formatting the output of `multidog()` to be
   a matrix.
@@ -92,7 +106,7 @@ This is a massive edit of the updog software. Major changes include:
     > discard all SNPs with a prop_mis **over 0.05**.
 - I've now exported some C++ functions that I think are useful. You
   can call them in the usual way:
-  <http://r-pkgs.had.co.nz/src.html#cpp-import>.
+  <https://r-pkgs.org/src.html#cpp-import>.
 
 
 # updog 0.99.0